+++ allocator.h
@@ -31,8 +31,8 @@
 
 void slab_init();
 void slab_cleanup(int quiet);
-void *slab_alloc(void *p, uint64_t size);
-void *slab_calloc(void *p, uint64_t items, uint64_t size);
+void *slab_alloc(void *p, size_t size);
+void *slab_calloc(void *p, size_t items, size_t size);
 void slab_free(void *p, void *address);
 void slab_release(void *p, void *address);
 int slab_cache_add(uint64_t size);
